+++ partials/front-page/above-the-fold.php
@@ -11,11 +11,7 @@ if ( $args['featured_posts_ids'] ) {
     return;
 }
 
-if ( $args['latest_news_posts_ids'] ) {
-    $latest_news_posts_ids = $args['latest_news_posts_ids'];
-} else {
-    return;
-}
+$latest_news_posts_ids = ! empty( $args['latest_news_posts_ids'] ) ? $args['latest_news_posts_ids'] : array();
 ?>
 <section class="front-page__above-the-fold container container--padded mt-2 mb-6 mb-s-5" data-testid="post-list">
   <div class="above-the-fold layout-grid">